app_files/requirements.txt
flask
transformers
torch
python-dotenv
accelerate